requirements.txt
langchain>=0.1.0
pymupdf>=1.23.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
streamlit>=1.30.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.18.0
python-dotenv>=1.1.0
transformers>=4.30.0